feat: render assignment totals as values are changed
