+++ app/components/staff_plan/assignment_totals_component.html.erb
@@ -1,0 +1,30 @@
+<td class="font-normal py-2 sm:pl-4 pl-0 pr-0 ml-1 sm:ml-0 w-1/2 sm:w-1/6" id="<%= dom_id(assignment, 'totals') %>">
+  <div
+    class="pl-6 sm:flex hidden flex-col max-w-fit space-y-5 cursor-pointer"
+  >
+    <div
+      class="sm:flex justify-between space-x-1 cursor-pointer relative pt-1"
+    >
+      <label
+        class="cursor-pointer text-sm leading-[18px] whitespace-nowrap text-right flex-shrink-0 min-w-10"
+      >Plan</label
+      ><span class="<%= sum_classes %>"
+    >
+      <%= planned_hours_sum %>
+      <span class="text-sm font-normal pl-1">hrs</span></span
+    >
+    </div>
+    <div
+      class="sm:flex justify-between space-x-1 cursor-pointer relative mt-[-7px]"
+    >
+      <label
+        class="cursor-pointer text-sm leading-[18px] whitespace-nowrap text-right flex-shrink-0 min-w-10"
+      >Actual</label
+      ><span class="<%= sum_classes %>"
+    >
+      <%= actual_hours_sum %>
+      <span class="text-sm font-normal pl-1">hrs</span></span
+    >
+    </div>
+  </div>
+</td>
